+++ superwall/src/main/java/com/superwall/sdk/paywall/presentation/internal/operators/LogErrors.kt
@@ -10,24 +10,26 @@ import com.superwall.sdk.paywall.presentation.internal.PaywallPresentationReques
 import com.superwall.sdk.paywall.presentation.internal.PaywallPresentationRequestStatusReason
 import com.superwall.sdk.paywall.presentation.internal.PresentationRequest
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
-fun Superwall.logErrors(
+suspend fun Superwall.logErrors(
     request: PresentationRequest,
     error: Throwable,
 ) {
     if (error is PaywallPresentationRequestStatusReason) {
-        GlobalScope.launch(Dispatchers.IO) {
-            val trackedEvent =
-                InternalSuperwallEvent.PresentationRequest(
-                    eventData = request.presentationInfo.eventData,
-                    type = request.flags.type,
-                    status = PaywallPresentationRequestStatus.NoPresentation,
-                    statusReason = error,
-                    factory = dependencyContainer,
-                )
-            track(trackedEvent)
+        withContext(Dispatchers.IO) {
+            launch {
+                val trackedEvent =
+                    InternalSuperwallEvent.PresentationRequest(
+                        eventData = request.presentationInfo.eventData,
+                        type = request.flags.type,
+                        status = PaywallPresentationRequestStatus.NoPresentation,
+                        statusReason = error,
+                        factory = dependencyContainer,
+                    )
+                track(trackedEvent)
+            }
         }
     }